Correct of mistakes 2
